Adding HydrateFallback
Hydration mismatch is a common issue in React applications. It occurs when the server-rendered HTML does not match the HTML generated by the client. This can cause unexpected behavior in your application, such as the loss of user input or the loss of scroll position. In Remix it can also cause FOUC (Flash of Unstyled Content).